requirements_server.txt
flask>=3.0.0
flask-socketio>=5.3.0
supabase>=2.3.0
sendgrid>=6.11.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
